+++ src/scaffold/root/backend/server.js
@@ -3,7 +3,6 @@
 
 const express = require('express');
 const http = require('http');
-const https = require('https');
 const fs = require('fs');
 const app = express();
 const config = require(process.cwd()+'/angular.json');