+++ src/scaffold/src/app/shared/components/home/home.component.ts
@@ -1,9 +1,8 @@
 import { Component, OnDestroy } from '@angular/core';
 import { trigger, state, style, transition, animate } from '@angular/animations';
 
-
 @Component({
-  selector: 'cmp-prefix1-app-home',
+  selector: 'app-home',
   templateUrl: 'home.component.html',
   styleUrls: ['home.component.css'],
   animations: [