+++ src/scaffold/src/app/shared/components/home/home.module.ts
@@ -1,15 +1,14 @@
-import { NgModule }           from '@angular/core';
-import { RouterModule }       from '@angular/router';
-import { CommonModule }  from '@angular/common';
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { CommonModule } from '@angular/common';
 
 import { HomeComponent } from './home.component';
-
 
 @NgModule({
   imports:      [ RouterModule,
                   CommonModule ],
-
-  declarations: [ HomeComponent ]
+  declarations: [ HomeComponent ],
+  exports: [ HomeComponent ]
 })
 
 export class HomeModule {}