+++ src/scaffold/src/app/shared/components/lazy/lazy.module.ts
@@ -1,9 +1,8 @@
-import { NgModule }           from '@angular/core';
-import { RouterModule }       from '@angular/router';
-import { CommonModule }  from '@angular/common';
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { CommonModule } from '@angular/common';
 
 import { LazyComponent } from './lazy.component';
-
 
 @NgModule({
   imports:      [ RouterModule,